requirements.txt
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.1
colorama==0.4.3
cycler==0.10.0
decorator==4.4.2
Django==2.2
django-cors-headers==3.2.1
django-filter==2.2.0
djangorestframework==3.11.0
h5py==2.10.0
idna==2.9
javascript==2015.32.59
kiwisolver==1.2.0
Markdown==3.2.1
matplotlib==3.2.1
monty==3.0.2
mpmath==1.1.0
networkx==2.4
numpy==1.18.2
palettable==3.3.0
pandas==1.0.3
Pillow==7.1.1
plotly==4.6.0
pydash==4.7.6
PyDispatcher==2.0.5
pyparsing==2.4.7
python-dateutil==2.8.1
python-vuejs==1.0.9
pytz==2019.3
requests==2.23.0
retrying==1.3.3
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
scipy==1.4.1
six==1.14.0
spglib==1.14.1.post0
sqlparse==0.3.1
sympy==1.5.1
tabulate==0.8.7
tqdm==4.45.0
urllib3==1.25.8
